Machine-Learning/Crop Scanner/cloud_function/requirements.txt
tensorflow == 2.12.0
numpy == 1.23.5
google-cloud-storage == 2.9.0
flask == 2.3.2
Pillow == 6.0.0